+++ src/basic/components/cart/CartList.tsx
@@ -1,0 +1,108 @@
+import { ShoppingBagIcon } from "../icons";
+import { CartItem } from "../../../types";
+import Button from "../ui/Button";
+import { XIcon } from "../icons";
+
+export default function CartList({
+  cart,
+  removeFromCart,
+  updateQuantity,
+  calculateItemTotal,
+}: {
+  cart: CartItem[];
+  removeFromCart: (productId: string) => void;
+  updateQuantity: (productId: string, quantity: number) => void;
+  calculateItemTotal: (item: CartItem) => number;
+}) {
+  if (cart.length === 0) {
+    return (
+      <div className="text-center py-8">
+        <ShoppingBagIcon
+          className="w-16 h-16 text-gray-300 mx-auto mb-4"
+          size="large"
+        />
+        <p className="text-gray-500 text-sm">장바구니가 비어있습니다</p>
+      </div>
+    );
+  }
+
+  return (
+    <div className="space-y-3">
+      {cart.map((item) => (
+        <CartItemCard
+          key={item.product.id}
+          item={item}
+          removeFromCart={removeFromCart}
+          updateQuantity={updateQuantity}
+          calculateItemTotal={calculateItemTotal}
+        />
+      ))}
+    </div>
+  );
+}
+
+const CartItemCard = ({
+  item,
+  removeFromCart,
+  updateQuantity,
+  calculateItemTotal,
+}: {
+  item: CartItem;
+  removeFromCart: (productId: string) => void;
+  updateQuantity: (productId: string, quantity: number) => void;
+  calculateItemTotal: (item: CartItem) => number;
+}) => {
+  const itemTotal = calculateItemTotal(item);
+  const originalPrice = item.product.price * item.quantity;
+  const hasDiscount = itemTotal < originalPrice;
+  const discountRate = hasDiscount
+    ? Math.round((1 - itemTotal / originalPrice) * 100)
+    : 0;
+
+  return (
+    <div key={item.product.id} className="border-b pb-3 last:border-b-0">
+      <div className="flex justify-between items-start mb-2">
+        <h4 className="text-sm font-medium text-gray-900 flex-1">
+          {item.product.name}
+        </h4>
+        <Button
+          onClick={() => removeFromCart(item.product.id)}
+          variant="ghost"
+          size="small"
+          className="text-gray-400 hover:text-red-500 ml-2 p-1 flex items-center justify-center"
+        >
+          <XIcon className="w-4 h-4" />
+        </Button>
+      </div>
+      <div className="flex items-center justify-between">
+        <div className="flex items-center">
+          <button
+            onClick={() => updateQuantity(item.product.id, item.quantity - 1)}
+            className="w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100"
+          >
+            <span className="text-xs">−</span>
+          </button>
+          <span className="mx-3 text-sm font-medium w-8 text-center">
+            {item.quantity}
+          </span>
+          <button
+            onClick={() => updateQuantity(item.product.id, item.quantity + 1)}
+            className="w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100"
+          >
+            <span className="text-xs">+</span>
+          </button>
+        </div>
+        <div className="text-right">
+          {hasDiscount && (
+            <span className="text-xs text-red-500 font-medium block">
+              -{discountRate}%
+            </span>
+          )}
+          <p className="text-sm font-medium text-gray-900">
+            {Math.round(itemTotal).toLocaleString()}원
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+};
